requirements.txt
Django==1.9
django-encrypted-fields==1.1.1
MySQL-python==1.2.5
pyasn1==0.1.9
pycrypto==2.6.1
python-keyczar==0.715
